Added specific .toString() to Filter.Compound - currently renders to e.g. 'Compound: !OR' (note again that negate str & spacer are hardcoded)
N.b. The operator is rendered as the internal (probably all-caps) form which may be undesirable (although it -is- localised, so a 'nicer' form could simply be achieved with localisation files). It also might be nice to add some logic to check if there are actually any Filter objects in the .filters array & append something like '(Empty)' or a count to the string.