Added .Year to Track.Sort.Release - asc/desc pair for sorting by a Track objs Release .year
N.b. Not notable in and of itself, but uses a new cleaner method to perform a .Desc sort - calls the .Asc() sort with an extra param set to 'true', which the asc uses to potentially invert the result of the a.year - b.year comparison, achieving both asc & desc with one set of logic. Other sorts should be refactored to use this style of logic.