Refactored DataCore->jsonSerialize() method to use ::$jsonFields/::getJsonFields() to generate the data (associative array) to be serialised. Also removed the (now redundant) ->jsonSerialize() overrides from SubDataCore, Release, Label & Track (the new system handles subclass specific fields automatically).
This completes the transition to the new more flexible serialisation system, which allows other classes/code to add (and technically redefine) fields to the serialised output of any DataCore based class.

N.b. This system is actually quite similar in a lot of ways to the new TAddMethod trait, and would probably actually make a pretty nifty trait itself (with a little work). One possible enhancement would be refactoring the way that the callbacks are processed to allow them access to the $this scope (and therefore allow them to access protected/private methods) like the TAddMethod trait does.